refactor the graph export/import
