+++ resotocore/resotocore/db/db_access.py
@@ -3,9 +3,7 @@ from abc import ABC
 from argparse import Namespace
 from datetime import datetime, timezone, timedelta
 from time import sleep
-from typing import Dict, List, Tuple, Union, cast, Optional, AsyncIterator
-from jsons import dumps, loads
-from attrs import frozen
+from typing import Dict, List, Tuple, Union, cast, Optional
 
 from arango import ArangoServerError
 from arango.client import ArangoClient
@@ -19,7 +17,7 @@ from resotocore.async_extensions import run_async
 from resotocore.core_config import CoreConfig
 from resotocore.db import SystemData
 from resotocore.db.arangodb_extensions import ArangoHTTPClient
-from resotocore.db.async_arangodb import AsyncArangoDB, AsyncCursor
+from resotocore.db.async_arangodb import AsyncArangoDB
 from resotocore.db.configdb import config_entity_db, config_validation_entity_db
 from resotocore.db.entitydb import EventEntityDb
 from resotocore.db.graphdb import ArangoGraphDB, GraphDB, EventGraphDB
@@ -34,7 +32,6 @@ from resotocore.error import NoSuchGraph, RequiredDependencyMissingError
 from resotocore.model.adjust_node import AdjustNode
 from resotocore.model.typed_model import from_js, to_js
 from resotocore.model.graph_access import EdgeTypes
-from resotocore.model.model import Kind
 from resotocore.types import Json
 from resotocore.ids import GraphName
 from resotocore.util import Periodic, utc, shutdown_process, uuid_str, check_graph_name
@@ -112,26 +109,6 @@ class DbAccess(ABC):
         await db.create_update_schema()
         return db
 
-    async def copy_graph(self, source: GraphName, destination: GraphName, validate_name: bool) -> GraphName:
-        """NOT THREAD SAFE! Be sure to use locks when calling this method."""
-        if validate_name:
-            check_graph_name(destination)
-
-        if not await self.db.has_graph(source):
-            raise ValueError(f"Source graph {source} does not exist")
-
-        source_db = self.get_graph_db(source, no_check=True)
-
-        await source_db.copy_graph(destination)
-
-        source_model_db = await self.get_graph_model_db(source)
-        destination_model_db = await self.get_graph_model_db(destination)
-
-        model_kinds = [kind async for kind in source_model_db.all()]
-        await destination_model_db.update_many(model_kinds)
-
-        return destination
-
     async def delete_graph(self, name: GraphName) -> None:
         def delete(name: GraphName) -> None:
             db = self.database
@@ -171,125 +148,6 @@ class DbAccess(ABC):
             event_db = EventGraphDB(graph_db, self.event_sender)
             self.graph_dbs[name] = event_db
             return event_db
-
-    async def export_graph(self, graph_name: GraphName) -> AsyncIterator[str]:
-        """
-        Export the graph as a newline-separated json object stream.
-        """
-        if not await self.db.has_graph(graph_name):
-            raise NoSuchGraph(graph_name)
-
-        graph = cast(EventGraphDB, self.get_graph_db(graph_name)).real
-        vertex_collection = graph_name
-        default_edge_collection = graph.edge_collection(EdgeTypes.default)
-        delete_edge_collection = graph.edge_collection(EdgeTypes.delete)
-        model_collection = self.graph_model_name(graph_name)
-
-        if not await self.db.has_collection(vertex_collection):
-            model_collection = "model"
-
-        async with self.db.begin_transaction(
-            read=[vertex_collection, default_edge_collection, delete_edge_collection, model_collection]
-        ) as tx:
-            # Snapshot isolation ensures that the counts are consistent with the cursor data
-            metadata = ExportMetadata(
-                serializer_version="0.1.0",
-                created_at=datetime.now().isoformat(),
-                model_collection_size=await tx.count(model_collection),
-                vertex_collection_size=await tx.count(vertex_collection),
-                default_edge_collection_size=await tx.count(default_edge_collection),
-                delete_edge_collection_size=await tx.count(delete_edge_collection),
-            )
-
-            # sturcture of the export file:
-            # 1. metadata
-            yield dumps(metadata)
-
-            # 2. model collection
-            cursor = AsyncCursor(await tx.all(model_collection), None)
-            async for doc in cursor:
-                yield dumps(doc)
-
-            # 3. vertex collection
-            cursor = AsyncCursor(await tx.all(vertex_collection), None)
-            async for doc in cursor:
-                yield dumps(doc)
-
-            # 4. default edge collection
-            cursor = AsyncCursor(await tx.all(default_edge_collection), None)
-            async for doc in cursor:
-                yield dumps(doc)
-
-            # 5. delete edge collection
-            cursor = AsyncCursor(await tx.all(delete_edge_collection), None)
-            async for doc in cursor:
-                yield dumps(doc)
-
-    async def import_graph(self, graph_name: GraphName, data: AsyncIterator[str]) -> None:
-        if await self.db.has_graph(graph_name):
-            raise ValueError(f"Graph {graph_name} already exists")
-
-        # temp graph to load the dump
-        temp_graph = cast(EventGraphDB, await self.create_graph(GraphName(graph_name + "_temp"))).real
-
-        metadata = from_js(loads(await data.__anext__()), ExportMetadata)
-        # check the serializer version, in the future we might need to support multiple versions
-        if metadata.serializer_version != "0.1.0":
-            raise ValueError(f"Unsupported dump version {metadata.serializer_version}")
-
-        # import the model directly to the target graph model collection
-        async def import_graph_model(data: AsyncIterator[str]) -> None:
-            if metadata.model_collection_size == 0:
-                return
-
-            position = 0
-            kinds: List[Json] = []
-            async for doc in data:
-                kinds.append(loads(doc))
-
-                # stop if we have reached the end of model
-                if position == metadata.model_collection_size - 1:
-                    break
-                position += 1
-
-            graph_model_db = await self.get_graph_model_db(graph_name)
-            await graph_model_db.update_many(from_js(kinds, List[Kind]))
-
-        # import the data into the temp graph
-        async def import_buffered(data: AsyncIterator[str], doc_num: int, collection: str) -> None:
-            if doc_num == 0:
-                return
-            position = 0
-            buffer = []
-            async for doc in data:
-                # collect a batch
-                buffer.append(loads(doc))
-                # insert the batch
-                if len(buffer) == 1000:
-                    await self.db.insert_many(collection, buffer)
-                    buffer = []
-
-                if position == doc_num - 1:
-                    break
-                position += 1
-
-            if len(buffer) > 0:
-                await self.db.insert_many(collection, buffer)
-
-        # step 1: import the graph model and the graph data into temporary collections
-        await import_graph_model(data)
-        await import_buffered(data, metadata.vertex_collection_size, temp_graph.vertex_name)
-        await import_buffered(
-            data, metadata.default_edge_collection_size, temp_graph.edge_collection(EdgeTypes.default)
-        )
-        await import_buffered(data, metadata.delete_edge_collection_size, temp_graph.edge_collection(EdgeTypes.delete))
-
-        # step 2: move the temporary graph collection to the final collection
-        # we're using the copy to do an atomic rename and rewrite the edge references
-        await self.copy_graph(temp_graph.vertex_name, graph_name, validate_name=False)
-
-        # step 3: delete the temporary graph
-        await self.delete_graph(temp_graph.name)
 
     def get_model_db(self) -> ModelDb:
         return self.model_db
@@ -418,13 +276,3 @@ class DbAccess(ABC):
         http_client = ArangoHTTPClient(args.graphdb_request_timeout, verify=verify)
         client = ArangoClient(hosts=args.graphdb_server, http_client=http_client)
         return client.db(args.graphdb_database, username=args.graphdb_username, password=args.graphdb_password)
-
-
-@frozen
-class ExportMetadata:
-    serializer_version: str
-    created_at: str
-    model_collection_size: int
-    vertex_collection_size: int
-    default_edge_collection_size: int
-    delete_edge_collection_size: int
